commands: move operator subcommands to subdirs
