+++ command/operator/autopilot/get/operator_autopilot_get.go
@@ -1,4 +1,4 @@
-package operautoget
+package get
 
 import (
 	"flag"
