+++ command/operator/autopilot/get/operator_autopilot_get_test.go
@@ -1,4 +1,4 @@
-package operautoget
+package get
 
 import (
 	"strings"
