+++ command/operator/autopilot/operator_autopilot_test.go
@@ -1,4 +1,4 @@
-package operauto
+package autopilot
 
 import (
 	"strings"
