+++ command/operator/autopilot/set/operator_autopilot_set_test.go
@@ -1,4 +1,4 @@
-package operautoset
+package set
 
 import (
 	"strings"
