+++ command/operator/raft/listpeers/operator_raft_list.go
@@ -1,4 +1,4 @@
-package operraftlist
+package listpeers
 
 import (
 	"flag"
