+++ command/operator/raft/listpeers/operator_raft_list_test.go
@@ -1,4 +1,4 @@
-package operraftlist
+package listpeers
 
 import (
 	"fmt"
