+++ command/operator/raft/operator_raft.go
@@ -1,4 +1,4 @@
-package operraft
+package raft
 
 import (
 	"github.com/mitchellh/cli"
