+++ command/operator/raft/operator_raft_test.go
@@ -1,4 +1,4 @@
-package operraft
+package raft
 
 import (
 	"strings"
