+++ command/operator/raft/removepeer/operator_raft_remove.go
@@ -1,4 +1,4 @@
-package operraftremove
+package removepeer
 
 import (
 	"flag"
