+++ command/operator/raft/removepeer/operator_raft_remove_test.go
@@ -1,4 +1,4 @@
-package operraftremove
+package removepeer
 
 import (
 	"strings"
